Moved authorization pages into authorizations folder
