+++ src/applications/representative-appoint/pages/authorizations/authorizeAddress.js
@@ -3,8 +3,8 @@ import {
   radioSchema,
   radioUI,
 } from '~/platform/forms-system/src/js/web-component-patterns';
-import { authorizationNote } from '../content/authorizeMedical';
-import { representativeTypeMap } from '../utilities/helpers';
+import { authorizationNote } from '../../content/authorizeMedical';
+import { representativeTypeMap } from '../../utilities/helpers';
 
 export const uiSchema = {
   'ui:description': ({ formData }) => {
