+++ src/applications/representative-appoint/pages/authorizations/authorizeMedicalSelect.js
@@ -3,7 +3,7 @@ import { checkboxGroupUI } from 'platform/forms-system/src/js/web-component-patt
 import {
   checkboxGroupSchemaWithReviewLabels,
   representativeTypeMap,
-} from '../utilities/helpers';
+} from '../../utilities/helpers';
 
 export const uiSchema = {
   'ui:description': ({ formData }) => {
